pyproject.toml
[project]
name = "repl-nix-hospitalinventorymanager"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pillow>=10.4.0",
    "plotly>=5.24.1",
    "psycopg2-binary>=2.9.9",
    "qrcode>=8.0",
    "streamlit-extras>=0.5.0",
    "streamlit>=1.39.0",
]
